feat: replace `Tabs` with `Sidebar`
