+++ app/routes/_app.alchemy._index/_index.tsx
@@ -1,5 +1,5 @@
 import * as RemixNode from '@remix-run/node'
 
 export function loader() {
-  return RemixNode.redirect('/alchemy/crafting')
+  return RemixNode.redirect('/alchemy/crafting/all')
 }
