+++ app/routes/_app.alchemy.crafting/_index.tsx
@@ -1,31 +1,20 @@
 import type * as RemixNode from '@remix-run/node'
 import * as RemixReact from '@remix-run/react'
-import Sidebar from '~/components/Sidebar'
 
 export const meta: RemixNode.MetaFunction = () => ({
   title: `Alchemy - Traveler Main`,
   description: `Alchemy page is to help you craft or convert items like in game`,
 })
 
-const navigation = [
-  { name: 'All', to: './all' },
-  { name: 'Enhancement', to: './enhancement' },
-  { name: 'Ascension', to: './ascension' },
-  { name: 'Talent', to: './talent' },
-]
-
 export default function CraftingLayout() {
   return (
-    <div className="py-8">
-      <div className="mx-auto max-w-3xl lg:grid lg:max-w-7xl lg:grid-cols-12 lg:gap-8">
-        <div className="col-span-12 lg:col-span-2">
-          <nav aria-label="Sidebar" className="sticky top-6">
-            <Sidebar navigation={navigation} />
-          </nav>
-        </div>
-        <main className="mt-8 lg:col-span-10 lg:mt-0">
-          <RemixReact.Outlet />
-        </main>
+    <div className="space-y-12">
+      <h1 className="text-2xl font-bold leading-7 text-gray-12 sm:truncate sm:text-3xl">
+        Alchemy Crafting
+      </h1>
+
+      <div className="mt-8 lg:col-span-10 lg:mt-0">
+        <RemixReact.Outlet />
       </div>
     </div>
   )
