+++ app/routes/_app.character.$name.manual-levelup/_index.tsx
@@ -78,7 +78,7 @@ export default function CharacterManualLevelupPage() {
   const busy = transition.state === 'submitting'
 
   return (
-    <div className="mt-8">
+    <div className="mt-12">
       <RemixReact.Form method="post">
         <div className="grid grid-cols-6 gap-6">
           <div className="col-span-6 sm:col-span-3">
