+++ app/routes/_app.character.$name.required-items/CharacterCustomFirstCell.tsx
