+++ app/routes/_app.character.$name.required-items/CustomTableHeading.tsx
