+++ app/routes/_app.character.$name.required-items/column.tsx
